requirements.txt
apache-airflow-providers-http
yfinance
pandas
SQLAlchemy
psycopg2-binary